Treesame-Commit-Id: 6c2f8e8c02997a3281bddd19d28b4f196804368e

+++ lib/cc/exporter/epub/files_directory.rb
@@ -7,6 +7,7 @@ module CC::Exporter::Epub
 
     def add_files
       files.each do |file_data|
+        next unless file_data[:exists]
         File.open(file_data[:path_to_file]) do |file|
           zip_file.add(file_data[:local_path], file)
         end
@@ -35,7 +36,7 @@ module CC::Exporter::Epub
     end
 
     def filename
-      @_filename ||= "#{SecureRandom.uuid}.zip"
+      @_filename ||= "#{Time.zone.now.strftime('%Y-%b-%d_%H-%M-%S')}.zip"
     end
 
     def zip_file
